requirements.txt
uvicorn==0.31.0
boto3==1.35.29
python-dotenv==1.0.1
fastapi==0.115.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
email-validator==2.2.0
httpx==0.27.2
python-multipart==0.0.11
PyJWT==2.9.0
openpyxl==3.1.5
